update leave balance

+++ database/migrations/2018_11_09_175624_create_leavebalances_table.php
@@ -19,7 +19,7 @@ class CreateLeavebalancesTable extends Migration
             $table->integer('leavetype_id');
             $table->date('start_date');
             $table->date('end_date');
-            $table->decimal('days',12,2);
+            $table->integer('days');
             $table->integer("creator_id");
             $table->timestamps();
         });
